fix subpath not deleted when using path pattern and add delete pvc test in pathpattern (#435)
* add delete pvc test in pathpattern

* add mount pod resource error test

+++ .github/scripts/test_case.py
@@ -538,6 +538,58 @@ def test_static_delete_pod():
     LOG.info("Test pass.")
 
 
+def test_pod_resource_err():
+    LOG.info("[test case] Pod resource error begin..")
+    # deploy pv
+    pv = PV(name="pv-resource-err", access_mode="ReadWriteMany", volume_handle="pv-resource-err",
+            secret_name=SECRET_NAME,
+            parameters={"juicefs/mount-cpu-request": "10", "juicefs/mount-memory-request": "50Gi",
+                        "juicefs/mount-cpu-limit": "10", "juicefs/mount-memory-limit": "50Gi", })
+    LOG.info("Deploy pv {}".format(pv.name))
+    pv.create()
+
+    # deploy pvc
+    pvc = PVC(name="pvc-resource-err", access_mode="ReadWriteMany", storage_name="", pv=pv.name)
+    LOG.info("Deploy pvc {}".format(pvc.name))
+    pvc.create()
+
+    # deploy pod
+    out_put = gen_random_string(6) + ".txt"
+    pod = Pod(name="app-resource-err", deployment_name="", replicas=1, namespace="default", pvc=pvc.name,
+              out_put=out_put)
+    pod.create()
+    LOG.info("Watch for pod {} for success.".format(pod.name))
+    result = pod.watch_for_success()
+    if not result:
+        raise Exception("Pods of deployment {} are not ready within 5 min.".format(pod.name))
+
+    # check mount point
+    LOG.info("Check mount point..")
+    volume_id = pvc.get_volume_id()
+    LOG.info("Get volume_id {}".format(volume_id))
+    result = check_mount_point(out_put)
+    if not result:
+        raise Exception("mount Point of /jfs/out.txt are not ready within 5 min.")
+
+    LOG.info("Check resources of mount pod..")
+    mount_pod = Pod(name=get_only_mount_pod_name(volume_id), deployment_name="", replicas=1, namespace=KUBE_SYSTEM)
+    spec = mount_pod.get_spec()
+    resource_requests = spec.containers[0].resources.requests
+    if resource_requests is not None and resource_requests["cpu"] != "" and resource_requests["memory"] != "":
+        raise Exception("Mount pod {} resources request is not none.".format(mount_pod.name))
+
+    # delete test resources
+    LOG.info("Remove pod {}".format(pod.name))
+    pod.delete()
+    LOG.info("Watch for pods for delete.".format(pod.name))
+    result = pod.watch_for_delete(1)
+    if not result:
+        raise Exception("Pods are not delete within 5 min.".format(pod.name))
+    LOG.info("Remove pvc {}".format(pvc.name))
+    pvc.delete()
+    LOG.info("Test pass.")
+
+
 def test_static_cache_clean_upon_umount():
     LOG.info("[test case] Pod with static storage and clean cache upon umount begin..")
     by_process = MOUNT_MODE == "process"
@@ -1162,6 +1214,81 @@ def test_path_pattern_in_storage_class():
         raise Exception("Pods of deployment {} are not delete within 5 min.".format(deployment.name))
     LOG.info("Remove pvc {}".format(pvc.name))
     pvc.delete()
+    LOG.info("Remove sc {}".format(pvc.name))
+    sc.delete()
+    return
+
+
+def test_dynamic_pvc_delete_with_path_pattern():
+    LOG.info("[test case] delete pvc with path pattern in storageClass begin..")
+    label_value = "def"
+    anno_value = "xyz"
+    # deploy sc
+    sc_name = "delete-pvc-path-pattern-dynamic"
+    sc = StorageClass(
+        name=sc_name, secret_name=SECRET_NAME,
+        parameters={"pathPattern": "${.PVC.namespace}-${.PVC.name}-${.PVC.labels.abc}-${.PVC.annotations.abc}"})
+    LOG.info("Deploy storageClass {}".format(sc.name))
+    sc.create()
+
+    # deploy pvc
+    pvc = PVC(name="delete-pvc-path-pattern-dynamic", access_mode="ReadWriteMany",
+              storage_name=sc.name, pv="", labels={"abc": label_value}, annotations={"abc": anno_value})
+    LOG.info("Deploy pvc {}".format(pvc.name))
+    pvc.create()
+
+    # deploy pod
+    out_put = gen_random_string(6) + ".txt"
+    deployment = Deployment(name="app-delete-pvc-path-pattern-dynamic", pvc=pvc.name, replicas=1, out_put=out_put)
+    LOG.info("Deploy deployment {}".format(deployment.name))
+    deployment.create()
+    pod = Pod(name="", deployment_name=deployment.name, replicas=deployment.replicas)
+    LOG.info("Watch for pods of {} for success.".format(deployment.name))
+    result = pod.watch_for_success()
+    if not result:
+        raise Exception("Pods of deployment {} are not ready within 10 min.".format(deployment.name))
+
+    # check mount point
+    LOG.info("Check mount point..")
+    check_path = "{}-{}-{}-{}/{}".format(KUBE_SYSTEM, pvc.name, label_value, anno_value, out_put)
+    result = check_mount_point(check_path)
+    if not result:
+        raise Exception("mount Point of {} are not ready within 5 min.".format(check_path))
+
+    LOG.info("Development delete..")
+    deployment.delete()
+    LOG.info("Watch deployment deleted..")
+    pod = Pod(name="", deployment_name=deployment.name, replicas=deployment.replicas)
+    LOG.info("Watch for pods of deployment {} for delete.".format(deployment.name))
+    result = pod.watch_for_delete(1)
+    if not result:
+        raise Exception("Pods of deployment {} are not delete within 5 min.".format(deployment.name))
+
+    LOG.info("PVC delete..")
+    pvc.delete()
+    for i in range(0, 60):
+        if pvc.check_is_deleted():
+            LOG.info("PVC is deleted.")
+            break
+        LOG.info("PVC is not deleted.")
+        time.sleep(5)
+
+    LOG.info("Check dir is deleted or not..")
+    file_exist = True
+    for i in range(0, 60):
+        f = pathlib.Path(GLOBAL_MOUNTPOINT + "/" + check_path)
+        if f.exists() is False:
+            file_exist = False
+            break
+        time.sleep(5)
+
+    if file_exist:
+        LOG.info("Mount point dir: ")
+        LOG.info(os.listdir(GLOBAL_MOUNTPOINT))
+        raise Exception("SubPath of volume_id {} still exists.".format(check_path))
+
+    LOG.info("Test pass.")
+    # delete test resources
     LOG.info("Remove sc {}".format(pvc.name))
     sc.delete()
     return
